day15 part1 progress
attempted a dfs using the code from a few days ago and was planning on simply finding all paths and calculating the sums of the risk levels. however, the number of total paths is n^n. my computer doesn't like that on the test input, and it won't like it on the real input. currently googling cheapest path algorithms as we speak